+++ lib/c-api/src/config.cc
@@ -7,572 +7,10 @@
 #include "core.hh"
 #include "lib.hh"
 
-#include <algorithm>
-#include <map>
-
 using namespace CppClingo::CAPI;
 
 namespace CppClingo::CAPI {
 namespace {
-
-class ClingoConfig {
-  public:
-    //! The type of configuration keys to refer to entries.
-    //!
-    //! Each configuration key is represented as unsigned integer.
-    using KeyType = Clasp::Cli::ClaspCliConfig::KeyType;
-    //! Type used for array indexing of configuration entries.
-    using IndexType = KeyType;
-    //! The available value types of configuration entries.
-    enum class ValueFlags : uint8_t {
-        none = 0,  //!< The entry cannot have a value.
-        get = 1,   //<! The entry can be read.
-        set = 2,   //!< The entry can be set.
-        value = 4, //!< The entry has a value.
-    };
-    CLINGO_ENABLE_BITSET_ENUM(ValueFlags, friend);
-
-    // A configuration entry interface to interact with configuration entries.
-    //
-    // This interface provides methods to check the value type of an entry, get and set values,
-    // and retrieve the size of array entries.
-    //
-    // Array handling is special. Configuration entries are arranged in a tree
-    // structure, where each path to a leave can have at most one array entry.
-    // If an entry has a parent that is an array, the index is passed to
-    // methods that require it.
-    class Entry {
-      public:
-        Entry() = default;
-        Entry(Entry &&other) = delete;
-        auto operator=(Entry &&other) -> Entry & = delete;
-        virtual ~Entry() = default;
-
-        //! Get information about the value of the entry.
-        [[nodiscard]] auto value_info(std::optional<IndexType> index) -> ValueFlags { return do_value_type(index); }
-        //! Get the value of the entry.
-        void get_value(std::optional<IndexType> index, std::string &value) { do_get_value(index, value); }
-        //! Set a new value for the entry.
-        void set_value(std::optional<KeyType> index, std::string_view value) { do_set_value(index, value); }
-        //! Returns the number of elements of an array entry.
-        //!
-        //! Returns std::nullopt if the entry is not an array.
-        auto size_array() -> std::optional<int> { return do_array_size(); }
-
-      private:
-        virtual auto do_value_type(std::optional<KeyType> index) -> ValueFlags = 0;
-        virtual void do_get_value(std::optional<KeyType> index, std::string &value) = 0;
-        virtual void do_set_value(std::optional<KeyType> index, std::string_view value) = 0;
-        virtual auto do_array_size() -> std::optional<int> = 0;
-    };
-    ClingoConfig(Clasp::Cli::ClaspCliConfig *config) : config_{config} {}
-
-    //! The root key of the configuration.
-    static auto key_root() -> KeyType { return Clasp::Cli::ClaspCliConfig::key_root; }
-    //! An invalid index for array keys.
-    //!
-    //! This index is used to mark that an array key is not associated with a
-    //! specific index.
-    static auto index_invalid() -> KeyType { return KeyType(-1); }
-
-    //! Retrieves information about the specified key.
-    //!
-    //! @note All out parameters are optional (i.e., can be null).
-    //!
-    //! @param key a handle to a key
-    //! @param n_children the number of subkeys for this key
-    //! @param array_info the length of array keys or -1 for non-array keys
-    //! @param value_info whether the key has a value, can be read, and can be set
-    //! @param help a description of the key
-    void key_info(KeyType key, int *n_children, int *array_info, ValueFlags *value_info, std::string *help) const {
-        init_(n_children, 0);
-        init_(array_info, -1);
-        init_(value_info, ValueFlags::none);
-        if (help != nullptr) {
-            help->clear();
-        }
-        if (auto clingo_key = is_clingo(key)) {
-            nodes_.at(clingo_key->key_id()).info(*clingo_key, n_children, array_info, value_info, help);
-        } else {
-            int n_values = 0;
-            if (auto ret = config_->getKeyInfo(key, n_children, array_info, help, &n_values); ret == -1) {
-                error_("invalid key");
-            }
-            if (value_info != nullptr) {
-                if (n_values < 0) {
-                    *value_info = ValueFlags::none;
-                } else if (n_values == 0) {
-                    *value_info = ValueFlags::get | ValueFlags::set;
-                } else {
-                    *value_info = ValueFlags::get | ValueFlags::set | ValueFlags::value;
-                }
-            }
-            // NOTE: the root key can have clingo subkeys
-            if (n_children != nullptr && key == key_root() && !nodes_.empty()) {
-                *n_children += nodes_.front().entries();
-            }
-        }
-    }
-
-    //! Get the key for an element at a specific index in an array entry.
-    //!
-    //! Returns the key corresponding to the element at the given index in the
-    //! array configuration entry specified by `key`. Throws if the entry is
-    //! not an array or the index is invalid.
-    //!
-    //! @param key key of the array entry
-    //! @param index index of the array element
-    //! @return key for the array element at the given index
-    [[nodiscard]] auto array_at(KeyType key, KeyType index) const -> KeyType {
-        if (auto clingo_key = is_clingo(key)) {
-            return nodes_.at(clingo_key->key_id()).array_at(*clingo_key, index).rep();
-        }
-        return check_(config_->getArrKey(key, index));
-    }
-
-    //! Get the key for a named subkey in a map entry.
-    //!
-    //! Returns the key corresponding to the subkey with the given name in the
-    //! map configuration entry specified by `key`. Throws if the subkey does
-    //! not exist.
-    //!
-    //! @param key key of the map entry
-    //! @param name name of the subkey
-    //! @return key for the named subkey
-    [[nodiscard]] auto map_at(KeyType key, std::string_view name) const -> KeyType {
-        if (auto clingo_key = is_clingo(key)) {
-            if (auto subkey = nodes_.at(clingo_key->key_id()).map_at(*clingo_key, name)) {
-                return subkey->rep();
-            }
-            error_("subkey not found");
-        }
-        if (key == key_root() && !nodes_.empty()) {
-            if (auto subkey = nodes_.front().map_at(Key{0}, name)) {
-                return subkey->rep();
-            }
-        }
-        return check_(config_->getKey(key, name));
-    }
-
-    //! Get the name of the nth subkey in a map entry.
-    //!
-    //! Returns the name of the subkey at the given index in the map
-    //! configuration entry specified by `key`. Throws if the index is out of
-    //! bounds.
-    //!
-    //! @param key key of the map entry
-    //! @param index index of the subkey
-    //! @return name of the nth subkey
-    [[nodiscard]] auto map_nth(KeyType key, uint32_t index) const -> std::string_view {
-        if (auto clingo_key = is_clingo(key)) {
-            if (auto name = nodes_.at(clingo_key->key_id()).map_nth(index)) {
-                return *name;
-            }
-            error_("index out of bounds");
-        }
-        if (key == key_root() && !nodes_.empty()) {
-            if (auto name = nodes_.front().map_nth(index)) {
-                return *name;
-            }
-        }
-        auto ret = config_->getSubkey(key, index);
-        if (ret.empty()) {
-            error_("index out of bounds");
-        }
-        return ret;
-    }
-
-    //! Get the value of a configuration entry.
-    //!
-    //! Returns the value associated with the given configuration key.
-    //! Throws if the key does not have a value or cannot be accessed.
-    //!
-    //! @param key key of the configuration entry
-    //! @return value of the configuration entry
-    [[nodiscard]] auto get_value(KeyType key) const -> std::string_view {
-        thread_local std::string value;
-        if (auto clingo_key = is_clingo(key)) {
-            nodes_.at(clingo_key->key_id()).get_value(clingo_key->index(), value);
-        } else {
-            if (config_->getValue(key, value) < 0) {
-                error_("cannot get value");
-            }
-        }
-        return value;
-    }
-
-    //! Set the value of a configuration entry.
-    //!
-    //! Sets the value associated with the given configuration key.
-    //! Throws if the key cannot be set or the value is invalid.
-    //!
-    //! @param key key of the configuration entry
-    //! @param value value to set
-    void set_value(KeyType key, std::string_view value) {
-        if (auto clingo_key = is_clingo(key)) {
-            nodes_.at(clingo_key->key_id()).set_value(clingo_key->index(), value);
-        } else {
-            if (config_->setValue(key, value) <= 0) {
-                error_("cannot set value");
-            }
-        }
-    }
-
-    //!  Adds a new clingo configuration entry to the configuration tree.
-    //!
-    //!  Validates the key name format and inserts the key as a child of the
-    //!  given parent key. The key name may include an optional index in the form
-    //!  "key[index]" to add an array key. The index of such a key is used as
-    //!  default index when the key is accessed without an index.
-    //!
-    //!  If an updater is provided, it will be associated with the key to handle
-    //!  changes to the key's value.
-    //!
-    //!  @param key parent clingo key under which the new key is added
-    //!  @param name name of the key
-    //!  @param description description string for the key
-    //!  @param entry optional entry to handle arrays and values
-    void add_entry(KeyType key, std::string_view name, std::string_view description,
-                   std::unique_ptr<Entry> entry = nullptr) {
-        if (nodes_.empty()) {
-            nodes_.emplace_back("the root key", nullptr);
-        }
-        auto clingo_key = Key{key == key_root() ? 0 : key};
-        auto dot = name.rfind('.');
-        if (dot == std::string_view::npos) {
-            dot = name.size();
-            clingo_key = parse_path_(clingo_key, name.substr(0, dot));
-            name = name.substr(dot + 1);
-        }
-
-        auto index = parse_name_(name);
-        auto &node = nodes_.at(clingo_key.key_id());
-        if (index && !entry) {
-            error_("array key without entry");
-        }
-        node.add_subkey(name, clingo_key, Key{static_cast<KeyType>(nodes_.size()), index});
-        nodes_.emplace_back(description, std::move(entry));
-    }
-
-  private:
-    static constexpr KeyType Bits = sizeof(KeyType) * 8;
-    static constexpr KeyType BitsIndex = 8;
-    static constexpr KeyType BitsKeyId = Bits - BitsIndex - 1;
-
-    static constexpr KeyType MaskClingo = KeyType{1} << (Bits - 1);
-    static constexpr KeyType MaskKeyId = ((KeyType{1} << BitsKeyId) - 1) << BitsIndex;
-    static constexpr KeyType MaskIndex = (KeyType{1} << BitsIndex) - 1;
-
-    static auto key_invalid() -> KeyType { return Clasp::Cli::ClaspCliConfig::key_invalid; }
-
-    struct FromRep {};
-    [[maybe_unused]] static constexpr FromRep from_rep{};
-
-    //! Checks if the given key is valid and returns it.
-    static auto check_(KeyType key) -> KeyType {
-        if (key == key_invalid()) {
-            error_("invalid key");
-        }
-        return key;
-    }
-
-    //! Helper to emit error messages for configuration issues.
-    template <class... T> static void error_(T const &...args) {
-        std::ostringstream oss;
-        oss << "configuration error: ";
-        (oss << ... << args); // NOLINT
-        throw std::runtime_error{oss.str()};
-    }
-
-    //! Helper to initialize a pointer with a default value if it is not null.
-    template <class P> static void init_(P *val, P def) {
-        if (val != nullptr) {
-            *val = def;
-        }
-    }
-
-    //! Helper to access the components of a configuration key.
-    //!
-    //! Keys are encoded as follows: Clingo Bit, Key Id Bits, Array Bits
-    //! Clingo Bit:
-    //! - A clingo key is invalid if this bit is not set.
-    //! Key Bits:
-    //! - The key id can use all available key id bits.
-    //! Array Bits:
-    //! - 0: Key is not on a path with an array key
-    //! - 1: Key is on a path with an array key that has no index (index_invalid())
-    //! - 2: Key is on a path with an array which has index `array bits - 2`
-    //! - A key is invalid if all array bits are set.
-    class Key {
-      public:
-        //! Constructs a new Key with the given key_id and optional index.
-        //!
-        //! Throws if the key cannot represent the given key_id or index.
-        explicit Key(KeyType key_id, std::optional<KeyType> index = std::nullopt)
-            : rep_{MaskClingo | encode_key_id(key_id) | encode_array(index)} {}
-
-        //! Constructs a Key from a raw representation.
-        //!
-        //! Throws if the representation does not represent a valid clingo key.
-        explicit Key([[maybe_unused]] FromRep tag, KeyType value) : rep_{value} {
-            if ((rep_ & MaskIndex) == MaskIndex || (rep_ & MaskClingo) == 0) {
-                error_("invalid key");
-            }
-        }
-
-        //! Get the id of the key.
-        [[nodiscard]] auto key_id() const -> KeyType { return (rep_ & MaskKeyId) >> BitsIndex; }
-        //! Get the index of the key.
-        [[nodiscard]] auto index() const -> std::optional<KeyType> {
-            auto res = rep_ & MaskIndex;
-            if (res == 0) {
-                return std::nullopt;
-            }
-            if (res == 1) {
-                return index_invalid();
-            }
-            return res - 2;
-        }
-        //! Transfer the array index of this key to the child key.
-        [[nodiscard]] auto subkey(Key child) const -> Key {
-            if (index() && child.index()) {
-                error_("multiple indices");
-            }
-            return index() ? Key{child.key_id(), index()} : child;
-        }
-        //! Get the raw representation of the key.
-        [[nodiscard]] auto rep() const -> KeyType { return rep_; }
-
-        //! Compare two keys for equality.
-        [[nodiscard, maybe_unused]] friend auto operator==(Key const &a, Key const &b) -> bool {
-            return a.rep_ == b.rep_;
-        }
-        //! Compare two keys.
-        [[nodiscard, maybe_unused]] friend auto operator<=>(Key const &a, Key const &b) -> std::strong_ordering {
-            return a.rep_ <=> b.rep_;
-        }
-
-      private:
-        static auto encode_array(std::optional<KeyType> index) -> KeyType {
-            KeyType rep = 0;
-            if (index) {
-                rep = 1;
-                if (*index != index_invalid()) {
-                    if (*index + 2 >= MaskIndex) {
-                        error_("index size exceeded");
-                    }
-                    rep = ((*index + 2) & MaskIndex);
-                }
-            }
-            return rep;
-        }
-        static auto encode_key_id(KeyType key_id) -> KeyType {
-            if (key_id >= (MaskKeyId >> BitsIndex)) {
-                error_("key size exceeded");
-            }
-            return key_id << BitsIndex;
-        }
-        KeyType rep_;
-    };
-
-    //! A concrete node in the configuration tree.
-    class Node {
-      public:
-        //! Constructs a new Node with the given description and optional updater.
-        explicit Node(std::string_view description, std::unique_ptr<Entry> entry)
-            : entry_{std::move(entry)}, description_{description} {}
-
-        //! The description of this node.
-        [[nodiscard]] auto description() const -> std::string_view { return description_; }
-        //! Get the number of entries of tihs node.
-        [[nodiscard]] auto entries() const -> int { return static_cast<int>(subkeys_.size()); }
-
-        //! Get the key for a subkey with the given name.
-        [[nodiscard]] auto map_at(Key key, std::string_view name) const -> std::optional<Key> {
-            auto it = subkeys_.find(name);
-            if (it != subkeys_.end()) {
-                return key.subkey(it->second);
-            }
-            return std::nullopt;
-        }
-        //! Get the name of the nth subkey in this node.
-        [[nodiscard]] auto map_nth(KeyType index) const -> std::optional<std::string_view> {
-            auto it = subkeys_.begin();
-            for (; index > 0 && it != subkeys_.end(); --index, ++it) {
-            }
-            return it != subkeys_.end() ? std::make_optional<std::string_view>(it->first) : std::nullopt;
-        }
-        //! Get the key for an array entry at the given index.
-        [[nodiscard]] auto array_at(Key key, KeyType index) const -> Key {
-            if (entry_ == nullptr || !entry_->size_array()) {
-                error_("not an array");
-            }
-            if (index == index_invalid()) {
-                error_("invalid index");
-            }
-            return Key{key.key_id(), index};
-        }
-        //! Get the value of this key.
-        void get_value(std::optional<KeyType> index, std::string &value) const {
-            if (entry_ == nullptr) {
-                error_("not a value");
-            }
-            if (index == index_invalid()) {
-                index = std::nullopt;
-            }
-            auto flags = entry_->value_info(index);
-            if (!intersects(flags, ValueFlags::get)) {
-                error_("cannot get value");
-            }
-            entry_->get_value(index, value);
-        }
-        //! Set the value of this key.
-        void set_value(std::optional<KeyType> index, std::string_view value) {
-            if (entry_ == nullptr) {
-                error_("not a value");
-            }
-            if (index == index_invalid()) {
-                index = std::nullopt;
-            }
-            auto flags = entry_->value_info(index);
-            if (!intersects(flags, ValueFlags::set)) {
-                error_("cannot set value");
-            }
-            entry_->set_value(index, value);
-        }
-        //! Get information about this key.
-        void info(Key key, int *n_children, int *array_info, ValueFlags *value_info, std::string *help) const {
-            if (n_children != nullptr) {
-                *n_children = static_cast<int>(subkeys_.size());
-            }
-            if (array_info != nullptr && entry_ != nullptr) {
-                *array_info = entry_->size_array().value_or(-1);
-            }
-            if (help != nullptr) {
-                auto desc = description();
-                help->assign(desc.begin(), desc.end());
-            }
-            if (value_info != nullptr && entry_ != nullptr) {
-                *value_info = entry_->value_info(key.index());
-            }
-        }
-
-        void add_subkey(std::string_view name, Key parent, Key child) {
-            if (parent.index() && child.index()) {
-                error_("multiple indices");
-            }
-            if (child.index()) {
-                child = Key{child.key_id(), index_invalid()};
-            }
-            subkeys_.emplace(name, child);
-        }
-
-      private:
-        std::unique_ptr<Entry> entry_;
-        std::map<std::string, Key, std::less<>> subkeys_;
-        std::string description_;
-    };
-
-    //! Parses a configuration key name, extracting an optional index if present.
-    //!
-    //! Accepts key names in the form "key" or "key[index]". If an index is
-    //! present, it is parsed and returned (where an empty index represents an
-    //! invalid index). The input name is modified to remove any index portion.
-    //!
-    //! Throws if the key is malformed.
-    //!
-    //! @param name[inout] reference to the key name
-    //! @return an optional index value
-    static auto parse_name_(std::string_view &name) -> std::optional<IndexType> {
-        auto res = std::optional<IndexType>{};
-        if (name.ends_with(']')) {
-            auto start = name.find_last_of('[');
-            if (start == std::string_view::npos) {
-                error_("missing '['");
-            }
-            auto index_str = name.substr(start + 1, name.size() - start - 2);
-            if (index_str.empty()) {
-                res = index_invalid();
-            } else {
-                res = 0;
-                auto [ptr, code] = std::from_chars(index_str.data(), index_str.data() + index_str.size(), *res);
-                if (code != std::errc{} && ptr != index_str.data() + index_str.size()) {
-                    error_("invalid index");
-                }
-                if (res.value() + 2 >= MaskIndex) {
-                    error_("index out of bounds");
-                }
-            }
-            name = name.substr(0, start);
-        }
-        if (name.empty()) {
-            error_("empty key");
-        }
-        if (std::isalpha(static_cast<unsigned char>(name.front())) == 0 && name.front() != '_') {
-            error_("invalid key");
-        }
-        if (!std::ranges::all_of(name,
-                                 [](char c) { return std::isalnum(static_cast<unsigned char>(c)) != 0 || c == '_'; })) {
-            error_("invalid key");
-        }
-        return res;
-    }
-
-    //! Resolves a clingo configuration key from a hierarchical path.
-    //!
-    //! Traverses the configuration tree starting from the given key and
-    //! following the path, which may contain dot-separated segments and
-    //! optional indices (e.g., "foo.bar[2].baz"). Returns a ClingoKey with the
-    //! resolved key id and index.
-    //!
-    //! The current implementation is limited to at most one index per key
-    //! segment.
-    //!
-    //! @param key The starting ClingoKey for resolution.
-    //! @param path The dot-separated configuration path to resolve.
-    //! @return The resolved ClingoKey.
-    //! @throws std::runtime_error if the path is invalid or resolution fails.
-    auto parse_path_(Key key, std::string_view path) -> Key {
-        for (size_t start = 0; start < path.size();) {
-            auto &cur = nodes_.at(key.key_id());
-
-            size_t dot = path.find('.', start);
-            auto last = dot == std::string_view::npos;
-            if (last) {
-                dot = path.size();
-            }
-            auto name = path.substr(start, dot - start);
-            auto index = parse_name_(name);
-            if (auto subkey = cur.map_at(key, name)) {
-                key = *subkey;
-            } else {
-                error_("key not found");
-            }
-            if (index) {
-                if (!key.index()) {
-                    error_("unexpected index");
-                }
-                key = Key{key.key_id(), index};
-            }
-            start = dot + 1;
-        }
-        return key;
-    }
-    // Checks if this key is a clasp key.
-    //
-    // Returns true if the upper bit is not set.
-    [[nodiscard]] static auto is_clasp(KeyType key) -> bool { return (key & MaskClingo) == 0; }
-
-    // Checks if this key is a clingo key.
-    //
-    // Returns true if the upper bit is set and the key is not the invalid key.
-    [[nodiscard]] static auto is_clingo(KeyType key) -> std::optional<Key> {
-        return !is_clasp(key) ? std::make_optional<Key>(key) : std::nullopt;
-    }
-
-    Clasp::Cli::ClaspCliConfig *config_;
-    std::vector<Node> nodes_;
-};
 
 inline auto cpp_cast(clingo_config_t const *config) -> Clasp::Cli::ClaspCliConfig const * {
     // NOLINTNEXTLINE
@@ -607,7 +45,9 @@ struct ConfigPrinter {
         int arr_len = 0;
         int vals = 0;
         cfg_->getKeyInfo(key, &map_keys, &arr_len, nullptr, &vals);
-        auto fi = [&, first = true]() mutable { return fill(std::exchange(first, false) ? first_indent : indent); };
+        auto fi = [&, first = true]() mutable {
+            return Util::fill(std::exchange(first, false) ? first_indent : indent);
+        };
         if (vals >= 0) {
             if (vals > 0) {
                 cfg_->getValue(key, val);
